REFACTOR: Changed MonthPill to ButtonPill
