+++ resources/ts/Pages/Courses/MonthlySection.tsx
@@ -1,5 +1,5 @@
 import Heading2 from "@/Components/Typography/Heading2";
-import MonthPill from "@/Components/Monthly/MonthPill";
+import ButtonPill from "@/Components/ButtonPill";
 import MonthlyOverview from "@/Components/Monthly/MonthlyOverview";
 import { useState } from "react";
 import { monthNames, seriesNames } from "@/constants";
@@ -28,7 +28,7 @@ export default function MonthlySection() {
                     <div role="list" className="grid grid-cols-3 grid-rows-4 gap-2 justify-items-stretch p-4">
                         {
                             monthNames.map((month, index) => (
-                                <MonthPill key={month} onPress={delayMonthlyOverview} isActive={selectedMonth === index} setActive={setSelectedMonth} idx={index} addClass={`w-full py-4`} >{month}</MonthPill>
+                                <ButtonPill key={month} onPress={delayMonthlyOverview} isActive={selectedMonth === index} setActive={setSelectedMonth} idx={index} addClass={`w-full py-4`} >{month}</ButtonPill>
                             ))
                         }
                     </div>
@@ -36,7 +36,7 @@ export default function MonthlySection() {
                     <div role="list" className="grid grid-cols-3 gap-2 p-4">
                         {
                             seriesNames.map((seriesElement, index) => (
-                                <MonthPill key={seriesElement.code} onPress={delayMonthlyOverview} isActive={selectedSeries === index} setActive={setSelectedSeries} idx={index} addClass={`w-full py-1`} >{seriesElement.name}</MonthPill>
+                                <ButtonPill key={seriesElement.code} onPress={delayMonthlyOverview} isActive={selectedSeries === index} setActive={setSelectedSeries} idx={index} addClass={`w-full py-1`} >{seriesElement.name}</ButtonPill>
                             ))
                         }
                     </div>
